+++ tests/elf_binary_printer_test.py
@@ -227,7 +227,6 @@ class ElfBinaryPrinterTests(BinaryPPrinterTest):
         Test printing a GTIRB that has no libraries with --dummy-so=yes
         """
         ir, _, _ = self.build_basic_ir()
-
         with self.binary_print(ir, "--dummy-so", "yes"):
             # Just verify binary_print succeeded.
             pass
@@ -236,10 +235,10 @@ class ElfBinaryPrinterTests(BinaryPPrinterTest):
         """
         Test printing version script
         """
-        ir, module, _ = self.build_basic_ir()
+        ir, module, text_bi = self.build_basic_ir()
 
-        proxy_foo = gth.add_proxy_block(module)
-        symbol_foo = gth.add_symbol(module, "foo", proxy_foo)
+        foo_block = gth.add_code_block(text_bi, b"\xC3")
+        symbol_foo = gth.add_symbol(module, "foo", foo_block)
         module.aux_data["elfSymbolInfo"].data[symbol_foo.uuid] = (
             0,
             "FUNC",
@@ -247,12 +246,21 @@ class ElfBinaryPrinterTests(BinaryPPrinterTest):
             "DEFAULT",
             0,
         )
-        proxy_bar = gth.add_proxy_block(module)
-        symbol_bar = gth.add_symbol(module, "bar", proxy_bar)
+        bar_block = gth.add_code_block(text_bi, b"\xC3")
+        symbol_bar = gth.add_symbol(module, "bar", bar_block)
         module.aux_data["elfSymbolInfo"].data[symbol_bar.uuid] = (
             0,
             "FUNC",
             "LOCAL",
+            "DEFAULT",
+            0,
+        )
+        proxy_baz = gth.add_proxy_block(module)
+        symbol_baz = gth.add_symbol(module, "baz", proxy_baz)
+        module.aux_data["elfSymbolInfo"].data[symbol_baz.uuid] = (
+            0,
+            "OBJECT",
+            "GLOBAL",
             "DEFAULT",
             0,
         )
@@ -266,22 +274,39 @@ class ElfBinaryPrinterTests(BinaryPPrinterTest):
                 # ElfSymVerDefs
                 {1: (["LIBA_1.0"], 0), 2: (["LIBA_2.0"], 0)},
                 # ElfSymVerNeeded
-                {"libmya.so": {1: "LIBA_1.0", 2: "LIBA_2.0"}},
+                {
+                    "libmya.so": {
+                        3: "LIBA_1.0",
+                        4: "LIBA_2.0",
+                    }
+                },
                 # ElfSymbolVersionsEntries
                 {
+                    # foo: global
                     symbol_foo.uuid: (1, False),
+                    # bar: local
                     symbol_bar.uuid: (2, False),
+                    # baz: external
+                    symbol_baz.uuid: (3, False),
                 },
             ),
         )
 
         vs = run_asm_pprinter_with_version_script(ir)
 
+        # Version LIBA_1 should not include the external symbol baz
         pattern1 = r"LIBA_1.0\s+{\s+global:\s+foo;\s+local:\s+\*;\s+};"
         self.assertRegexMatch(vs, pattern1)
+        self.assertTrue("baz" not in vs)
+
+        # Version LIBA_2 should not print out the local symbol bar
         pattern2 = r"LIBA_2.0\s+{\s+local:\s+\*;\s+};"
         self.assertRegexMatch(vs, pattern2)
         self.assertTrue("bar" not in vs)
+
+        with self.binary_print(ir, "--dummy-so", "yes", "--shared"):
+            # Just verify binary_print succeeded.
+            pass
 
     def test_base_version(self):
         """