+++ CHANGELOG.md
@@ -12,9 +12,14 @@
   * Blacklist `_GLOBAL_OFFSET_TABLE_` symbol from dummy-so libraries.
   * Explicitly link `ld-linux*.so` if `-nodefaultlibs` is enabled.
   * Support GOT, PAGE and GOT, OFST symbolic expression attributes for split .got loads on MIPS.
+  * Populate non-local symbols under `global:` and add `local: *;` for local
+    symbols in version-script.
+  * Change version scripts that only use the module's symbol version
+    definitions (not its needed symbols).
+  * Create a different version script for dummy-so dependencies using the
+    needed symbol versions.
   * Populate non-local symbols under `global:` and add `local: *;` for local symbols in version-script
   * For ambiguous symbols, choose one symbol to keep original symbol name.
-  * For symbols with base version, only print out the connector and do not print out the base version in .symver directives.
 
 # 2.1.0
   * `--asm` option now prints the assembly for each module of an IR separately